+++ contracts/modules/delegates/IexecPocoBoostDelegate.sol
@@ -234,7 +234,9 @@ contract IexecPocoBoostDelegate is IexecPocoBoost, IexecAccessorsBoost, Delegate
         }
         deal.workerReward = 0; // TODO: Update and test
         deal.beneficiary = _requestorder.beneficiary;
-        deal.deadline = 0; // TODO: Update and test
+        deal.deadline = (block.timestamp +
+            m_categories[_requestorder.category].workClockTimeRef *
+            CONTRIBUTION_DEADLINE_RATIO).toUint48();
         deal.botSize = volume.toUint24();
         deal.tag = vars.tag;
         deal.callback = _requestorder.callback;
@@ -288,6 +290,12 @@ contract IexecPocoBoostDelegate is IexecPocoBoost, IexecAccessorsBoost, Delegate
     ) external {
         IexecLibCore_v5.DealBoost storage deal = m_dealsBoost[dealId];
         bytes32 taskId = keccak256(abi.encodePacked(dealId, index));
+        //TODO: Add `claimBoost(..)`
+        require(
+            m_tasks[taskId].status == IexecLibCore_v5.TaskStatusEnum.UNSET,
+            "PocoBoost: Task status not unset"
+        );
+        require(block.timestamp < deal.deadline, "PocoBoost: Deadline reached");
         // TODO: Check enclave challenge if TEE bit set
         // Check scheduler signature
         require(
@@ -311,6 +319,15 @@ contract IexecPocoBoostDelegate is IexecPocoBoost, IexecAccessorsBoost, Delegate
             "PocoBoost: Invalid enclave signature"
         );
 
+        /**
+         * @dev Prevent reentrancy before external call
+         * Minimal reuse of Poco Classic task map.
+         * Benefit: Fetching task status is unchanged for clients
+         */
+        m_tasks[taskId].status = IexecLibCore_v5.TaskStatusEnum.COMPLETED;
+
+        emit ResultPushedBoost(dealId, index, results);
+
         if (target != address(0)) {
             require(resultsCallback.length > 0, "PocoBoost: Callback requires data");
             (bool success, ) = target.call{gas: m_callbackgas}(
@@ -319,7 +336,6 @@ contract IexecPocoBoostDelegate is IexecPocoBoost, IexecAccessorsBoost, Delegate
             success; // silent unused variable warning
             require(gasleft() > m_callbackgas / 63, "PocoBoost: Not enough gas after callback");
         }
-        emit ResultPushedBoost(dealId, index, results);
     }
 
     /**